Add boggle_oo.js script
This is our in-progress solution for the boggle word-finder.  Our
current version (boggle.js) crashes if there are multiple instances
of the first letters of a word and some of those instances do not
result in the final word.

The goal of this feature is to take in a word and return ALL coordinates
for the word that exist on the board.  This version returns some but not
all of those instances.

This version is not integrated with the game_board and instead only
runs on a text version of the game.  We will need to fully integrate it
with the game once it is working.