Add new constant: INPRODUCTION
this variable is currently set to false

It is currently only used in connection with whether words can be
boggled while the game is paused.  In production, this practice should
not be allowed.  However, it is very useful for testing.

We are trying to keep it quiet that the game has a pause feature in the
first place.